add logout, blocklist

+++ resources/product.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.product import ProductModel
 
 
@@ -25,10 +25,11 @@ class Product(Resource):
         item=ProductModel.find_by_name(name)
         if item:
             return item.json()
-        return {'message': 'Product could not be found'}, 404
+        return {'message': 'Product could not be found unfortunately'}, 404
 
 
-
+    
+    @jwt_required(fresh=True)
     def post(self, name):
         
         if ProductModel.find_by_name(name):
@@ -48,6 +49,9 @@ class Product(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims=get_jwt()
+        if not claims['is_admin']:
+            return{'message': 'Admin privilege required'}
         product=ProductModel.find_by_name(name)
         if product:
             product.delete_from_db()
@@ -74,5 +78,13 @@ class Product(Resource):
 
 
 class ProductList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return{'products':[product.json() for product in ProductModel.find_all()]}
+        user_id=get_jwt_identity()
+        products = [product.json() for product in ProductModel.find_all()]
+        if user_id:
+            return {'products': products}, 200
+        return{
+            'products':[product['name'] for product in products], 
+            'message': 'More data will be available if you log in'
+        }, 200
